+++ example/json_rest_server_example.dart
@@ -1,6 +1,0 @@
-import 'package:json_rest_server/json_rest_server.dart';
-
-void main() {
-  var awesome = Awesome();
-  print('awesome: ${awesome.isAwesome}');
-}