+++ lib/json_rest_server.dart
@@ -1,8 +1,0 @@
-/// Support for doing something awesome.
-///
-/// More dartdocs go here.
-library json_rest_server;
-
-export 'src/json_rest_server_base.dart';
-
-// TODO: Export any libraries intended for clients of this package.