+++ CHANGELOG.md
@@ -5,3 +5,19 @@
 ## 1.0.2
 
 - Beta version.
+
+## 1.0.3
+
+- Bug fix create project
+
+## 1.0.4
+
+- Bug fix logs
+
+## 1.0.5
+
+- adding shortcuts to run (json_rest_client, jsonRestClient or jsr)
+
+## 1.0.6
+
+- Fix Readme and example